@@ -5,25 +5,26 @@
   "dependencies": {
     "@commitlint/cli": "^17.4.4",
     "@commitlint/config-conventional": "^17.4.4",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^26.0.15",
     "@types/lodash": "^4.14.176",
     "@types/node": "^12.0.0",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@types/react-router-dom": "^5.3.2",
+    "@types/react": "^18.0.28",
+    "@types/react-dom": "^18.0.11",
+    "@types/react-router-dom": "^5.3.3",
     "@types/styled-components": "^5.1.15",
     "husky": "^8.0.3",
     "lodash": "^4.17.21",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "4.0.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.8.1",
+    "react-scripts": "5.0.1",
+    "semantic-release": "^20.1.0",
     "styled-components": "^5.3.1",
     "typescript": "4.4.4",
-    "web-vitals": "^1.0.1"
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -31,7 +32,8 @@
     "test": "react-scripts test",
     "lint": "eslint src --ext .js,.jsx,.ts,.tsx",
     "eject": "react-scripts eject",
-    "prepare": "husky install"
+    "prepare": "husky install",
+    "semantic-release": "semantic-release"
   },
   "eslintConfig": {
     "extends": [
@@ -73,10 +75,11 @@
     ]
   },
   "engines": {
-    "npm": "8.19.2",
-    "node": "16.18.1"
+    "npm": "9.5.0",
+    "node": "18.14.2"
   },
-  "devDependencies": {
-    "pinst": "^3.0.0"
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/fang-lin/game-of-life.git"
   }
 }
