@@ -1,22 +1,22 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import {GlobalStyle} from './index.styles';
 import {App} from './App';
 // import reportWebVitals from './reportWebVitals';
-import {HashRouter, Switch, Redirect, Route} from 'react-router-dom';
+import {HashRouter, Routes, Navigate, Route} from 'react-router-dom';
 import {combinePathToURL, defaultParams, routerPath, stringifyParams} from './App.functions';
 
-ReactDOM.render(
+createRoot(document.getElementById('root')!).render(
     <React.StrictMode>
         <HashRouter>
-            <Switch>
-                <Route path={routerPath()} component={App}/>
-                <Redirect to={combinePathToURL(stringifyParams(defaultParams))}/>
-            </Switch>
+            <Routes>
+                <Route path={routerPath()} element={<App/>}/>
+                <Route path="*" element={<Navigate to={combinePathToURL(stringifyParams(defaultParams))} replace={true}/>}/>
+                {/*<Navigate to={combinePathToURL(stringifyParams(defaultParams))}/>*/}
+            </Routes>
         </HashRouter>
         <GlobalStyle/>
     </React.StrictMode>,
-    document.getElementById('root')
 );
 
 // If you want to start measuring performance in your app, pass a function
